Temporary disable failed C# unit tests
work-items: #52447
git-tfs-id: [https://sede-pt-ce2.visualstudio.com/DefaultCollection]$/Basin Modeling/IBS/Trunk;C54022

+++ development/libraries/casaAPI/src/CauldronApp.h
@@ -87,7 +87,7 @@ namespace casa
 
       /// @brief Get run time limits for this application [Minutes], a job scheduler is using this number
       /// @return minutes number for the run time limitation
-      int runTimeLimit() { return m_runTimeLim; }
+      size_t runTimeLimit() { return m_runTimeLim; }
 
       /// @brief Add application option
       /// @param opt new option
